op move Navigator into lib
