+++ app/src/main/java/com/myth/mama/MainViewModel.kt
@@ -2,9 +2,9 @@ package com.myth.mama
 
 import android.os.Bundle
 import com.myth.arch.mvvm2.MythViewModel
-import com.myth.mama.viewmodelext.startActivity
-import com.myth.mama.viewmodelext.startActivityForResult
-import com.myth.mama.viewmodelext.toast
+import com.myth.arch.mvvm2.ext.startActivity
+import com.myth.arch.mvvm2.ext.startActivityForResult
+import com.myth.arch.mvvm2.ext.toast
 
 class MainViewModel : MythViewModel() {
