--- app/src/main/java/com/myth/mama/viewmodelext/ToastExt.kt
@@ -1,8 +1,0 @@
-package com.myth.mama.viewmodelext
-
-import com.myth.arch.mvvm2.MythViewModel
-import com.myth.arch.mvvm2.ext.ToastExt
-
-fun MythViewModel.toast(text: String) {
-    ToastExt.toast(this, text)
-}
