+++ myth-arch-mvvm-lib/src/main/java/com/myth/arch/mvvm2/ext/NavigateExt.kt
@@ -1,4 +1,4 @@
-package com.myth.mama.viewmodelext
+package com.myth.arch.mvvm2.ext
 
 import android.app.Activity
 import android.content.Intent
@@ -7,7 +7,6 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import com.myth.arch.mvvm2.MythView
 import com.myth.arch.mvvm2.MythViewModel
-import com.myth.arch.mvvm2.ext.MythViewModelExt
 import kotlin.reflect.KClass
 
 class NavigateExt : MythViewModelExt<KClass<Activity>>() {
@@ -129,7 +128,10 @@ fun <T : Activity> MythViewModel.startActivity(cls: KClass<T>) {
 }
 
 fun <T : Activity> MythViewModel.startActivity(cls: KClass<T>, data: Bundle) {
-    Navigate2Ext.startActivity(this, Navigate2Data(cls, data))
+    Navigate2Ext.startActivity(
+        this,
+        Navigate2Data(cls, data)
+    )
 }
 
 fun <T : Activity> MythViewModel.startActivityForResult(
@@ -137,5 +139,8 @@ fun <T : Activity> MythViewModel.startActivityForResult(
     data: Bundle,
     requestCode: Int
 ) {
-    Navigate3Ext.startActivityForResult(this, Navigate3Data(cls, data, requestCode))
+    Navigate3Ext.startActivityForResult(
+        this,
+        Navigate3Data(cls, data, requestCode)
+    )
 }
