+++ myth-arch-mvvm-lib/src/main/java/com/myth/arch/mvvm2/ext/ToastExt.kt
@@ -24,3 +24,7 @@ class ToastExt : MythViewModelExt<String>() {
         })
     }
 }
+
+fun MythViewModel.toast(text: String) {
+    ToastExt.toast(this, text)
+}
